# src/backend/CompanyHero.Api/Contracts/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "CompanyHero API",
    "description": "Verbindlicher Vertrag nach A-009 und K13: Kennungen und Dezimalwerte als Strings, Zeitpunkte mit Offset, Null und fehlendes Feld unterscheidbar.",
    "version": "1"
  },
  "paths": {
...
    "/api/access/join-codes": {
      "get": {
        "tags": [
          "CompanyHero.Api"
        ],
        "operationId": "ListJoinCodes",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/JoinCodeResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "CompanyHero.Api"
        ],
        "operationId": "CreateJoinCode",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateJoinCodeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JoinCodeResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/access/join-codes/{id}/revoke": {
      "post": {
        "tags": [
          "CompanyHero.Api"
        ],
        "operationId": "RevokeJoinCode",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/access/role-codes": {
      "post": {
        "tags": [
          "CompanyHero.Api"
        ],
        "operationId": "IssueRoleCode",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IssueRoleCodeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleCodeResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/access/role-codes/redeem": {
      "post": {
        "tags": [
          "CompanyHero.Api"
        ],
        "operationId": "RedeemRoleCode",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RedeemRoleCodeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/access/policy": {
      "get": {
        "tags": [
          "CompanyHero.Api"
        ],
        "operationId": "GetLoginPolicy",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginPolicyResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "CompanyHero.Api"
        ],
        "operationId": "UpdateLoginPolicy",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginPolicyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginPolicyResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/access/policy/preview": {
      "post": {
        "tags": [
          "CompanyHero.Api"
        ],
        "operationId": "PreviewLoginPolicy",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginPolicyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginPolicyImpactResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/access/providers": {
      "get": {
        "tags": [
          "CompanyHero.Api"
        ],
        "operationId": "ListTenantProviders",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TenantProviderResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "CompanyHero.Api"
        ],
        "operationId": "ConfigureTenantProvider",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfigureProviderRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantProviderResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/access/providers/{id}": {
      "delete": {
        "tags": [
          "CompanyHero.Api"
        ],
        "operationId": "DisableTenantProvider",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/access/kiosk-devices": {
      "get": {
        "tags": [
          "CompanyHero.Api"
        ],
        "operationId": "ListKioskDevices",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/KioskDeviceAdminResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "CompanyHero.Api"
        ],
        "operationId": "CreateKioskDevice",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateKioskDeviceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KioskDeviceAdminResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/access/kiosk-devices/{id}/revoke": {
      "post": {
        "tags": [
          "CompanyHero.Api"
        ],
        "operationId": "RevokeKioskDevice",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ConfigureProviderRequest": {
        "required": [
          "displayName",
          "issuer",
          "clientId",
          "clientSecret"
        ],
        "type": "object",
        "properties": {
          "displayName": {
            "type": "string"
          },
          "issuer": {
            "type": "string"
          },
          "clientId": {
            "type": "string"
          },
          "clientSecret": {
            "type": "string"
          }
        }
      },
...
      "CreateJoinCodeRequest": {
        "required": [
          "usageLimit",
          "validDays"
        ],
        "type": "object",
        "properties": {
          "usageLimit": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "validDays": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          }
        }
      },
      "CreateKioskDeviceRequest": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          }
        }
      },
...
      "HttpValidationProblemDetails": {
        "required": [
          "errors"
        ],
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        }
      },
      "IssueRoleCodeRequest": {
        "required": [
          "role",
          "email"
        ],
        "type": "object",
        "properties": {
          "role": {
            "type": "string"
          },
          "email": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "JoinCodeResponse": {
        "required": [
          "joinCodeId",
          "code",
          "createdAt",
          "expiresAt",
          "usageLimit",
          "usedCount",
          "revokedAt",
          "joinUrl"
        ],
        "type": "object",
        "properties": {
          "joinCodeId": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "usageLimit": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "usedCount": {
            "type": "integer",
            "format": "int32"
          },
          "revokedAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "joinUrl": {
            "type": "string"
          }
        }
      },
...
      "KioskDeviceAdminResponse": {
        "required": [
          "deviceId",
          "name",
          "createdAt",
          "registeredAt",
          "revokedAt",
          "lastSeenAt",
          "loginCount",
          "registrationCode",
          "registrationExpiresAt"
        ],
        "type": "object",
        "properties": {
          "deviceId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "registeredAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "revokedAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "lastSeenAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "loginCount": {
            "type": "integer",
            "format": "int32"
          },
          "registrationCode": {
            "type": [
              "null",
              "string"
            ]
          },
          "registrationExpiresAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          }
        }
      },
...
      "LoginPolicyImpactResponse": {
        "required": [
          "personsWithoutWay"
        ],
        "type": "object",
        "properties": {
          "personsWithoutWay": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "LoginPolicyRequest": {
        "required": [
          "magicLink",
          "passkey",
          "kiosk",
          "disabledProviderKeys",
          "forcedProviderKeys",
          "kioskIdleSeconds"
        ],
        "type": "object",
        "properties": {
          "magicLink": {
            "type": "boolean"
          },
          "passkey": {
            "type": "boolean"
          },
          "kiosk": {
            "type": "boolean"
          },
          "disabledProviderKeys": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "forcedProviderKeys": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "kioskIdleSeconds": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "LoginPolicyResponse": {
        "required": [
          "magicLink",
          "passkey",
          "kiosk",
          "disabledProviderKeys",
          "forcedProviderKeys",
          "kioskIdleSeconds",
          "availableProviders"
        ],
        "type": "object",
        "properties": {
          "magicLink": {
            "type": "boolean"
          },
          "passkey": {
            "type": "boolean"
          },
          "kiosk": {
            "type": "boolean"
          },
          "disabledProviderKeys": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "forcedProviderKeys": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "kioskIdleSeconds": {
            "type": "integer",
            "format": "int32"
          },
          "availableProviders": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProviderResponse"
            }
          }
        }
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "ProviderResponse": {
        "required": [
          "key",
          "displayName"
        ],
        "type": "object",
        "properties": {
          "key": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          }
        }
      },
...
      "RedeemRoleCodeRequest": {
        "required": [
          "code",
          "realName"
        ],
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "realName": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "RoleCodeResponse": {
        "required": [
          "roleCodeId",
          "code",
          "role",
          "expiresAt"
        ],
        "type": "object",
        "properties": {
          "roleCodeId": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "TenantProviderResponse": {
        "required": [
          "providerId",
          "key",
          "displayName",
          "issuer",
          "clientId",
          "validatedAt",
          "disabledAt"
        ],
        "type": "object",
        "properties": {
          "providerId": {
            "type": "string"
          },
          "key": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "issuer": {
            "type": "string"
          },
          "clientId": {
            "type": "string"
          },
          "validatedAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "disabledAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          }
        }
      },
...
    }
  },
...
}